Twelfth Commit - Refraction Again
So refractive indices need to be 1.0 or above to not flip the image.
Also looks like using refraction and reflection in the same material
isn't possible. The reflection I have implemented now is technically 2
sharks being rendered on top of on another and therefore completely
wrong. I just wanted to see if it would work. Looking into raytracing.