Twenty-first Commit - Format Changes
started ordering things the way Professor Pang wants them in the
eventual submit folder. The report html (index.html in the report
folder) is just a placeholder and is the proposal for the moment. Added
screenshots and a README. Looked at the new models, some of them seem
broken. We may want to consider taking them out. Also, took out the
floor (because we weren't using it in reflection/refraction
calculations).